requirements.txt
PySocks=1.7.1
